gallium: use unsigned long to silence warnings with 64-bit build

+++ src/mesa/state_tracker/st_cb_texture.c
@@ -205,7 +205,7 @@ __memcpy(void *to, const void *from, size_t n)
 static void *
 do_memcpy(void *dest, const void *src, size_t n)
 {
-   if ((((unsigned) src) & 63) || (((unsigned) dest) & 63)) {
+   if ((((unsigned long) src) & 63) || (((unsigned long) dest) & 63)) {
       return __memcpy(dest, src, n);
    }
    else
